chore: bump sqldelight to 2.0.0 (#2045)
* chore: bump sqldelight to 2.0.0

* replace sourceFolders with srcDirs

* fix sqldelight breaking changes

* remove unwanted changes

* detekt

* fix: ios driver changes

* detekt

+++ persistence/src/appleMain/kotlin/com/wire/kalium/persistence/db/GlobalDatabaseProvider.kt
@@ -48,17 +48,16 @@ actual class GlobalDatabaseProvider(
 
     init {
         NSFileManager.defaultManager.createDirectoryAtPath(storePath, true, null, null)
-
         val schema = GlobalDatabase.Schema
         val driver = NativeSqliteDriver(
             DatabaseConfiguration(
                 name = dbName,
-                version = schema.version,
+                version = schema.version.toInt(),
                 create = { connection ->
                     wrapConnection(connection) { schema.create(it) }
                 },
                 upgrade = { connection, oldVersion, newVersion ->
-                    wrapConnection(connection) { schema.migrate(it, oldVersion, newVersion) }
+                    wrapConnection(connection) { schema.migrate(it, oldVersion.toLong(), newVersion.toLong()) }
                 },
                 extendedConfig = DatabaseConfiguration.Extended(
                     basePath = storePath
